Use correct type.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@13416 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ utils/cre/CREWrapperFormulae.cpp
@@ -49,7 +49,7 @@ int CREWrapperFormulae::experience() const
 QStringList CREWrapperFormulae::archs() const
 {
     QStringList archs;
-    for (int i = 0; i < myFormulae->arch_names; i++)
+    for (size_t i = 0; i < myFormulae->arch_names; i++)
     {
         archs.append(myFormulae->arch_name[i]);
     }
